BridgeGuard_AI-main/requirements.txt
torch
torchvision
timm
numpy
opencv-python
Pillow
pytorch-grad-cam